Minor changes to eventlog -- added event log image.
The rest are changes to the final AddLinks() call on most all pages -- cosmetic -- but I think it's clearer.

Addition of calls to eventlog for quite a few other things. One major one is missing localization and theme items. They are just logged and replaced with their [NAME] instead of throwing an exception. Hopefully this will make adding theme and text items to future version LESS PAINFUL.