Updated the look/feel of Error.aspx and added a feature to show error messsages specified (in one case, ForumPage redirects to it).
When I turned YAF to release there were some problems... so those got fixed. Added a "YOU ARE IN DEBUG MODE" message to the footer so we don't have to answer the same question: "How do I get rid of the performance SQL data at the bottom of the screen. etc."