Cleaned up deletemessage a bit -- made it a bit clearer and started using the MessagePost control.
Added recalculation of NumPosts to message_delete SQL as the numposts is incredibly out of sync.

Localization for "Deleted by User/Mod" text for a deleted message.

Added some validation to editforum.

Added 1 to page size for posts on the page which has "Sponser" text... need to improve for "deleted" messages.