Did it right this time:
ForumRoot is the EXTERNAL forum root (set with BaseUrl).
ForumFileRoot is the INTERNAL forum root (set with Root).

Made the necessary changes to the forum so there is a DIFFERENCE in the code. Now the external URL can be different than the internal one.

Added fix for security issues associated with Avatar data without ContentTypes.

Modified Mail system according to "WESclyburnw" -- SMTP is pulled from mail.config is NO LONGER in board settings.

Mail system no longer uses a thread -- it now uses a timer. Added a Queue function to the SendMail.