Fix for issue  ID: 2930531 on SF:
"

In 1.9.4 RC1, in a user (tried on admin) creates a new topic with any title
but enters only a '.' (dot) as message content, this will cause an
exception upon posting.
To make matters serious, this exception is then also thrown upon every
access to the forum this topic is in, effectively blocking the forum. It
can be resolved by manually changing the post content in database, even
changing to another single letter is enough."