Added proper Logging system. YafSendMailThread now uses DI as do all the background tasks.
All pages and controls now support injection and have an instance of ILogger available.

From now on, do not make calls to eventlog in the db. Instead, use: this.Logger

You should not just do this.Get<ILogger>() as ILogger is from a provider and uses the creating Type to log that information.