Wayne Dupree:
"Here’s a quick fix for a problem with the mobile user agent checker. I found that some user agents (like the one for the Motorola Droid) we’re triggering the mobile version of the forum and found that the “toLower” had been left off of the userAgent string. Small fix but it made all the mobile browsers work correctly again for me."