Ok here is my last new Big feature for the next release. Single Sign On via Facebook (Needs to be enabled in the Host Settings).
It is now possible to login via Facebook for Security Check an existing User needs to enter the Facebook User Id in the Profile Settings and activate the SSO Feature. When a User try to log in via facebook, yaf checks if a user exists with that email address

There is also a Setting which auto Register Users that are not exists.