The Current Points System is now Converted in to a Reputation System instead of Adding Points (Reputation) automatically when Users Posts a Message, a User can give Reputation to another User (once a Day).
A New Setting can turn on/off the Reputation System "Use Reputation System".

There is a Setting "Min. Value to allow Add Reputation" - Check if the User matches minimal requirements for voting up. (Default Value is: 1)

There is a Setting "Min. Value to allow Remove Reputation:" - Check if the User matches minimal requirements for voting down. (Default Value is: 100)

There is a Setting "Allow Negative Reputation" - If allowed Users can get Negative Reputation Numbers.

And there are two Settings to Define a Min Negative Value (0%) and a Maximum Value (100%) to Generate a Progressbar that Shows the Users Karma.