A bit of a hack for speed but changed the email address need for avatars to Lazy<> so that it's not strickly required.
It was causing the topic pages to call providers for the user's email address on every user... which is not needed unless we're dealing with gravatars.

In that situation code will still need to be added to cache the email address.

The next step is a user repository abstraction that will allow for cleaner user access anyway.