Moved Typing to IDbFunctionSession so it can optimized/improved.
Converted Models to straight POCO objects. Added Active and ActiveAccess
models and repositories plus updated the system.

Got rid of IDBBroker interface altogether. Totally wrong headed there...

Added Value Injector project as a CORE dependancy... it now handles mapping
of IDataReaders -- and it's so-so fast. Ultimately, probably want to generate
some cached IL for speed. But for now, it's fine.