new spam bot prevention system
it is already in test mode in the support forum. It is as "Spam Word"
system you define a word (including regex syntax).

For example "kitchen.*co.uk". This catches all kitchen spammers!!

What it does:

- During Registration it checks the Homepage URL Profile field (usually
used by the kitchen spammers) against the spam words => If any word
matches, it uses the "Handling of Bots during Registration" in the host
settings of what to do with that user

- When a user Edits his profile and edits the the Homepage URL Profile
it will be checked against the spam words => If any word matches, it
will only log the "spam" detection (for now)

- When a user Edits his signature  it will be checked against the spam
words => If any word matches, it will only log the "spam" detection (for
now)

- When a user posts a message in the forum it checks the message against
the spam words if "Choose a SPAM Service for Automatically SPAM
Checking" in the host settings is set to "Internal Spam Check Words"
=> If any word matches, it will use the action set in the "Handling of
as SPAM Detected Messages" host setting